% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administration_api.R
\docType{class}
\name{AdministrationApi}
\alias{AdministrationApi}
\title{Administration operations}
\format{
An \code{R6Class} generator object
}
\description{
AdministrationApi
}
\details{
Mednet EDC API

The Mednet EDC (Electronic Data Capture) REST API provides a single-point of access for reading data stored across iMednet data services. This specification documents all available endpoints in the latest version of the API. 

The version of the OpenAPI document: 1.0.15
Generated by: https://openapi-generator.tech
}
\examples{
\dontrun{
####################  ListUsers  ####################

library(openapi)
var_study_key <- "study_key_example" # character | Study key identifying the study context for the request
var_page <- 0 # integer | Page index to retrieve (0-based) (Optional)
var_size <- 25 # integer | Number of items to return per page (max 500) (Optional)
var_sort <- "sort_example" # character | Sorting criteria in the format `property,ASC` or `property,DESC`. Can be repeated. (Optional)
var_include_inactive <- FALSE # character | For user listing, whether to include inactive users (Optional)

#List users and their roles in a study
api_instance <- AdministrationApi$new()

# Configure API key authorization: apiKeyAuth
api_instance$api_client$api_keys["x-api-key"] <- Sys.getenv("API_KEY")

# Configure API key authorization: securityKeyAuth
api_instance$api_client$api_keys["x-imn-security-key"] <- Sys.getenv("API_KEY")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$ListUsers(var_study_key, page = var_page, size = var_size, sort = var_sort, include_inactive = var_include_inactivedata_file = "result.txt")
result <- api_instance$ListUsers(var_study_key, page = var_page, size = var_size, sort = var_sort, include_inactive = var_include_inactive)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AdministrationApi-new}{\code{AdministrationApi$new()}}
\item \href{#method-AdministrationApi-ListUsers}{\code{AdministrationApi$ListUsers()}}
\item \href{#method-AdministrationApi-ListUsersWithHttpInfo}{\code{AdministrationApi$ListUsersWithHttpInfo()}}
\item \href{#method-AdministrationApi-clone}{\code{AdministrationApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdministrationApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-AdministrationApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new AdministrationApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdministrationApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdministrationApi-ListUsers"></a>}}
\if{latex}{\out{\hypertarget{method-AdministrationApi-ListUsers}{}}}
\subsection{Method \code{ListUsers()}}{
List users and their roles in a study
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdministrationApi$ListUsers(
  study_key,
  page = 0,
  size = 25,
  sort = NULL,
  include_inactive = FALSE,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study_key}}{Study key identifying the study context for the request}

\item{\code{page}}{(optional) Page index to retrieve (0-based) (default value: 0)}

\item{\code{size}}{(optional) Number of items to return per page (max 500) (default value: 25)}

\item{\code{sort}}{(optional) Sorting criteria in the format `property,ASC` or `property,DESC`. Can be repeated.}

\item{\code{include_inactive}}{(optional) For user listing, whether to include inactive users (default value: FALSE)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
UserList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdministrationApi-ListUsersWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AdministrationApi-ListUsersWithHttpInfo}{}}}
\subsection{Method \code{ListUsersWithHttpInfo()}}{
List users and their roles in a study
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdministrationApi$ListUsersWithHttpInfo(
  study_key,
  page = 0,
  size = 25,
  sort = NULL,
  include_inactive = FALSE,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study_key}}{Study key identifying the study context for the request}

\item{\code{page}}{(optional) Page index to retrieve (0-based) (default value: 0)}

\item{\code{size}}{(optional) Number of items to return per page (max 500) (default value: 25)}

\item{\code{sort}}{(optional) Sorting criteria in the format `property,ASC` or `property,DESC`. Can be repeated.}

\item{\code{include_inactive}}{(optional) For user listing, whether to include inactive users (default value: FALSE)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (UserList) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdministrationApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AdministrationApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdministrationApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
