% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnyType.R
\docType{class}
\name{AnyType}
\alias{AnyType}
\title{AnyType}
\description{
Represents an arbitrary object with key/value pairs.
}
\details{
AnyType Class

Generic container for flexible JSON objects used in the API.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnyType-new}{\code{AnyType$new()}}
\item \href{#method-AnyType-toJSON}{\code{AnyType$toJSON()}}
\item \href{#method-AnyType-toList}{\code{AnyType$toList()}}
\item \href{#method-AnyType-toSimpleType}{\code{AnyType$toSimpleType()}}
\item \href{#method-AnyType-fromJSON}{\code{AnyType$fromJSON()}}
\item \href{#method-AnyType-fromJSONString}{\code{AnyType$fromJSONString()}}
\item \href{#method-AnyType-toJSONString}{\code{AnyType$toJSONString()}}
\item \href{#method-AnyType-validateJSON}{\code{AnyType$validateJSON()}}
\item \href{#method-AnyType-toString}{\code{AnyType$toString()}}
\item \href{#method-AnyType-isValid}{\code{AnyType$isValid()}}
\item \href{#method-AnyType-getInvalidFields}{\code{AnyType$getInvalidFields()}}
\item \href{#method-AnyType-print}{\code{AnyType$print()}}
\item \href{#method-AnyType-clone}{\code{AnyType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyType-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnyType-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyType$new(additional_properties = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnyType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyType$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyType-toList"></a>}}
\if{latex}{\out{\hypertarget{method-AnyType-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyType$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyType-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-AnyType-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyType$toSimpleType()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnyType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyType$fromJSON(input_json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AnyType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyType$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AnyType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyType$toJSONString(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyType-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnyType-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyType$validateJSON(input)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyType-toString"></a>}}
\if{latex}{\out{\hypertarget{method-AnyType-toString}{}}}
\subsection{Method \code{toString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyType$toString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyType-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-AnyType-isValid}{}}}
\subsection{Method \code{isValid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyType$isValid()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyType-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-AnyType-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyType$getInvalidFields()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyType-print"></a>}}
\if{latex}{\out{\hypertarget{method-AnyType-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyType$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnyType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
