% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding_list.R
\docType{class}
\name{CodingList}
\alias{CodingList}
\title{CodingList}
\format{
An \code{R6Class} generator object
}
\description{
CodingList Class

CodingList Class
}
\details{
Create a new CodingList
}
\examples{

## ------------------------------------------------
## Method `CodingList$toList`
## ------------------------------------------------

# convert array of CodingList (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{\link{Metadata} [optional]}

\item{\code{pagination}}{\link{Pagination} [optional]}

\item{\code{data}}{list(\link{Coding}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CodingList-new}{\code{CodingList$new()}}
\item \href{#method-CodingList-toJSON}{\code{CodingList$toJSON()}}
\item \href{#method-CodingList-toList}{\code{CodingList$toList()}}
\item \href{#method-CodingList-toSimpleType}{\code{CodingList$toSimpleType()}}
\item \href{#method-CodingList-fromJSON}{\code{CodingList$fromJSON()}}
\item \href{#method-CodingList-toJSONString}{\code{CodingList$toJSONString()}}
\item \href{#method-CodingList-fromJSONString}{\code{CodingList$fromJSONString()}}
\item \href{#method-CodingList-validateJSON}{\code{CodingList$validateJSON()}}
\item \href{#method-CodingList-toString}{\code{CodingList$toString()}}
\item \href{#method-CodingList-isValid}{\code{CodingList$isValid()}}
\item \href{#method-CodingList-getInvalidFields}{\code{CodingList$getInvalidFields()}}
\item \href{#method-CodingList-print}{\code{CodingList$print()}}
\item \href{#method-CodingList-clone}{\code{CodingList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CodingList-new"></a>}}
\if{latex}{\out{\hypertarget{method-CodingList-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new CodingList class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CodingList$new(metadata = NULL, pagination = NULL, data = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{metadata}

\item{\code{pagination}}{pagination}

\item{\code{data}}{data}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CodingList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CodingList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CodingList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CodingList-toList"></a>}}
\if{latex}{\out{\hypertarget{method-CodingList-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CodingList$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CodingList as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of CodingList (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CodingList-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-CodingList-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert CodingList to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CodingList$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CodingList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CodingList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of CodingList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CodingList$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CodingList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CodingList-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CodingList-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CodingList$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
CodingList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CodingList-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CodingList-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of CodingList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CodingList$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CodingList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CodingList-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CodingList-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to CodingList and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CodingList$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CodingList-toString"></a>}}
\if{latex}{\out{\hypertarget{method-CodingList-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CodingList$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of CodingList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CodingList-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-CodingList-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CodingList$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CodingList-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-CodingList-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CodingList$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CodingList-print"></a>}}
\if{latex}{\out{\hypertarget{method-CodingList-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CodingList$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CodingList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CodingList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CodingList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
