% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components_schemas_interval_forms_item.R
\docType{class}
\name{ComponentsSchemasIntervalFormsItem}
\alias{ComponentsSchemasIntervalFormsItem}
\title{ComponentsSchemasIntervalFormsItem}
\format{
An \code{R6Class} generator object
}
\description{
ComponentsSchemasIntervalFormsItem Class

ComponentsSchemasIntervalFormsItem Class
}
\details{
Create a new ComponentsSchemasIntervalFormsItem
}
\examples{

## ------------------------------------------------
## Method `ComponentsSchemasIntervalFormsItem$toList`
## ------------------------------------------------

# convert array of ComponentsSchemasIntervalFormsItem (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{formId}}{Form ID scheduled in the interval integer [optional]}

\item{\code{formKey}}{Form key scheduled in the interval character [optional]}

\item{\code{formName}}{Form name scheduled in the interval character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ComponentsSchemasIntervalFormsItem-new}{\code{ComponentsSchemasIntervalFormsItem$new()}}
\item \href{#method-ComponentsSchemasIntervalFormsItem-toJSON}{\code{ComponentsSchemasIntervalFormsItem$toJSON()}}
\item \href{#method-ComponentsSchemasIntervalFormsItem-toList}{\code{ComponentsSchemasIntervalFormsItem$toList()}}
\item \href{#method-ComponentsSchemasIntervalFormsItem-toSimpleType}{\code{ComponentsSchemasIntervalFormsItem$toSimpleType()}}
\item \href{#method-ComponentsSchemasIntervalFormsItem-fromJSON}{\code{ComponentsSchemasIntervalFormsItem$fromJSON()}}
\item \href{#method-ComponentsSchemasIntervalFormsItem-toJSONString}{\code{ComponentsSchemasIntervalFormsItem$toJSONString()}}
\item \href{#method-ComponentsSchemasIntervalFormsItem-fromJSONString}{\code{ComponentsSchemasIntervalFormsItem$fromJSONString()}}
\item \href{#method-ComponentsSchemasIntervalFormsItem-validateJSON}{\code{ComponentsSchemasIntervalFormsItem$validateJSON()}}
\item \href{#method-ComponentsSchemasIntervalFormsItem-toString}{\code{ComponentsSchemasIntervalFormsItem$toString()}}
\item \href{#method-ComponentsSchemasIntervalFormsItem-isValid}{\code{ComponentsSchemasIntervalFormsItem$isValid()}}
\item \href{#method-ComponentsSchemasIntervalFormsItem-getInvalidFields}{\code{ComponentsSchemasIntervalFormsItem$getInvalidFields()}}
\item \href{#method-ComponentsSchemasIntervalFormsItem-print}{\code{ComponentsSchemasIntervalFormsItem$print()}}
\item \href{#method-ComponentsSchemasIntervalFormsItem-clone}{\code{ComponentsSchemasIntervalFormsItem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentsSchemasIntervalFormsItem-new"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentsSchemasIntervalFormsItem-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ComponentsSchemasIntervalFormsItem class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentsSchemasIntervalFormsItem$new(
  formId = NULL,
  formKey = NULL,
  formName = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formId}}{Form ID scheduled in the interval}

\item{\code{formKey}}{Form key scheduled in the interval}

\item{\code{formName}}{Form name scheduled in the interval}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentsSchemasIntervalFormsItem-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentsSchemasIntervalFormsItem-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentsSchemasIntervalFormsItem$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentsSchemasIntervalFormsItem-toList"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentsSchemasIntervalFormsItem-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentsSchemasIntervalFormsItem$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ComponentsSchemasIntervalFormsItem as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of ComponentsSchemasIntervalFormsItem (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentsSchemasIntervalFormsItem-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentsSchemasIntervalFormsItem-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert ComponentsSchemasIntervalFormsItem to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentsSchemasIntervalFormsItem$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentsSchemasIntervalFormsItem-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentsSchemasIntervalFormsItem-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ComponentsSchemasIntervalFormsItem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentsSchemasIntervalFormsItem$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ComponentsSchemasIntervalFormsItem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentsSchemasIntervalFormsItem-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentsSchemasIntervalFormsItem-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentsSchemasIntervalFormsItem$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ComponentsSchemasIntervalFormsItem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentsSchemasIntervalFormsItem-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentsSchemasIntervalFormsItem-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ComponentsSchemasIntervalFormsItem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentsSchemasIntervalFormsItem$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ComponentsSchemasIntervalFormsItem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentsSchemasIntervalFormsItem-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentsSchemasIntervalFormsItem-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ComponentsSchemasIntervalFormsItem and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentsSchemasIntervalFormsItem$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentsSchemasIntervalFormsItem-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentsSchemasIntervalFormsItem-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentsSchemasIntervalFormsItem$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ComponentsSchemasIntervalFormsItem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentsSchemasIntervalFormsItem-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentsSchemasIntervalFormsItem-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentsSchemasIntervalFormsItem$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentsSchemasIntervalFormsItem-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentsSchemasIntervalFormsItem-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentsSchemasIntervalFormsItem$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentsSchemasIntervalFormsItem-print"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentsSchemasIntervalFormsItem-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentsSchemasIntervalFormsItem$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComponentsSchemasIntervalFormsItem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ComponentsSchemasIntervalFormsItem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComponentsSchemasIntervalFormsItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
