% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_object.R
\docType{class}
\name{InlineObject}
\alias{InlineObject}
\title{InlineObject}
\format{
An \code{R6Class} generator object
}
\description{
InlineObject Class

InlineObject Class
}
\details{
Create a new InlineObject
}
\examples{

## ------------------------------------------------
## Method `InlineObject$toList`
## ------------------------------------------------

# convert array of InlineObject (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{\link{Metadata} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InlineObject-new}{\code{InlineObject$new()}}
\item \href{#method-InlineObject-toJSON}{\code{InlineObject$toJSON()}}
\item \href{#method-InlineObject-toList}{\code{InlineObject$toList()}}
\item \href{#method-InlineObject-toSimpleType}{\code{InlineObject$toSimpleType()}}
\item \href{#method-InlineObject-fromJSON}{\code{InlineObject$fromJSON()}}
\item \href{#method-InlineObject-toJSONString}{\code{InlineObject$toJSONString()}}
\item \href{#method-InlineObject-fromJSONString}{\code{InlineObject$fromJSONString()}}
\item \href{#method-InlineObject-validateJSON}{\code{InlineObject$validateJSON()}}
\item \href{#method-InlineObject-toString}{\code{InlineObject$toString()}}
\item \href{#method-InlineObject-isValid}{\code{InlineObject$isValid()}}
\item \href{#method-InlineObject-getInvalidFields}{\code{InlineObject$getInvalidFields()}}
\item \href{#method-InlineObject-print}{\code{InlineObject$print()}}
\item \href{#method-InlineObject-clone}{\code{InlineObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new InlineObject class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject$new(metadata = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{metadata}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject-toList"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InlineObject as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of InlineObject (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert InlineObject to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of InlineObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InlineObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
InlineObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of InlineObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InlineObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to InlineObject and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject-toString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of InlineObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
