% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_object4.R
\docType{class}
\name{InlineObject4}
\alias{InlineObject4}
\title{InlineObject4}
\format{
An \code{R6Class} generator object
}
\description{
InlineObject4 Class

InlineObject4 Class
}
\details{
Create a new InlineObject4
}
\examples{

## ------------------------------------------------
## Method `InlineObject4$toList`
## ------------------------------------------------

# convert array of InlineObject4 (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{\link{Metadata} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InlineObject4-new}{\code{InlineObject4$new()}}
\item \href{#method-InlineObject4-toJSON}{\code{InlineObject4$toJSON()}}
\item \href{#method-InlineObject4-toList}{\code{InlineObject4$toList()}}
\item \href{#method-InlineObject4-toSimpleType}{\code{InlineObject4$toSimpleType()}}
\item \href{#method-InlineObject4-fromJSON}{\code{InlineObject4$fromJSON()}}
\item \href{#method-InlineObject4-toJSONString}{\code{InlineObject4$toJSONString()}}
\item \href{#method-InlineObject4-fromJSONString}{\code{InlineObject4$fromJSONString()}}
\item \href{#method-InlineObject4-validateJSON}{\code{InlineObject4$validateJSON()}}
\item \href{#method-InlineObject4-toString}{\code{InlineObject4$toString()}}
\item \href{#method-InlineObject4-isValid}{\code{InlineObject4$isValid()}}
\item \href{#method-InlineObject4-getInvalidFields}{\code{InlineObject4$getInvalidFields()}}
\item \href{#method-InlineObject4-print}{\code{InlineObject4$print()}}
\item \href{#method-InlineObject4-clone}{\code{InlineObject4$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject4-new"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject4-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new InlineObject4 class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject4$new(metadata = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{metadata}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject4-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject4-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject4$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject4-toList"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject4-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject4$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InlineObject4 as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of InlineObject4 (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject4-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject4-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert InlineObject4 to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject4$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject4-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject4-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of InlineObject4
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject4$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InlineObject4
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject4-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject4-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject4$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
InlineObject4 in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject4-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject4-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of InlineObject4
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject4$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InlineObject4
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject4-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject4-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to InlineObject4 and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject4$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject4-toString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject4-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject4$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of InlineObject4
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject4-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject4-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject4$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject4-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject4-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject4$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject4-print"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject4-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject4$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject4-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject4-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject4$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
