% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_object5.R
\docType{class}
\name{InlineObject5}
\alias{InlineObject5}
\title{InlineObject5}
\format{
An \code{R6Class} generator object
}
\description{
InlineObject5 Class

InlineObject5 Class
}
\details{
Create a new InlineObject5
}
\examples{

## ------------------------------------------------
## Method `InlineObject5$toList`
## ------------------------------------------------

# convert array of InlineObject5 (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{\link{Metadata} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InlineObject5-new}{\code{InlineObject5$new()}}
\item \href{#method-InlineObject5-toJSON}{\code{InlineObject5$toJSON()}}
\item \href{#method-InlineObject5-toList}{\code{InlineObject5$toList()}}
\item \href{#method-InlineObject5-toSimpleType}{\code{InlineObject5$toSimpleType()}}
\item \href{#method-InlineObject5-fromJSON}{\code{InlineObject5$fromJSON()}}
\item \href{#method-InlineObject5-toJSONString}{\code{InlineObject5$toJSONString()}}
\item \href{#method-InlineObject5-fromJSONString}{\code{InlineObject5$fromJSONString()}}
\item \href{#method-InlineObject5-validateJSON}{\code{InlineObject5$validateJSON()}}
\item \href{#method-InlineObject5-toString}{\code{InlineObject5$toString()}}
\item \href{#method-InlineObject5-isValid}{\code{InlineObject5$isValid()}}
\item \href{#method-InlineObject5-getInvalidFields}{\code{InlineObject5$getInvalidFields()}}
\item \href{#method-InlineObject5-print}{\code{InlineObject5$print()}}
\item \href{#method-InlineObject5-clone}{\code{InlineObject5$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject5-new"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject5-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new InlineObject5 class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject5$new(metadata = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{metadata}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject5-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject5-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject5$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject5-toList"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject5-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject5$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InlineObject5 as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of InlineObject5 (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject5-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject5-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert InlineObject5 to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject5$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject5-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject5-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of InlineObject5
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject5$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InlineObject5
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject5-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject5-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject5$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
InlineObject5 in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject5-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject5-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of InlineObject5
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject5$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InlineObject5
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject5-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject5-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to InlineObject5 and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject5$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject5-toString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject5-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject5$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of InlineObject5
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject5-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject5-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject5$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject5-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject5-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject5$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject5-print"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject5-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject5$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineObject5-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InlineObject5-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineObject5$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
