% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_list.R
\docType{class}
\name{IntervalList}
\alias{IntervalList}
\title{IntervalList}
\format{
An \code{R6Class} generator object
}
\description{
IntervalList Class

IntervalList Class
}
\details{
Create a new IntervalList
}
\examples{

## ------------------------------------------------
## Method `IntervalList$toList`
## ------------------------------------------------

# convert array of IntervalList (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{\link{Metadata} [optional]}

\item{\code{pagination}}{\link{Pagination} [optional]}

\item{\code{data}}{list(\link{Interval}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IntervalList-new}{\code{IntervalList$new()}}
\item \href{#method-IntervalList-toJSON}{\code{IntervalList$toJSON()}}
\item \href{#method-IntervalList-toList}{\code{IntervalList$toList()}}
\item \href{#method-IntervalList-toSimpleType}{\code{IntervalList$toSimpleType()}}
\item \href{#method-IntervalList-fromJSON}{\code{IntervalList$fromJSON()}}
\item \href{#method-IntervalList-toJSONString}{\code{IntervalList$toJSONString()}}
\item \href{#method-IntervalList-fromJSONString}{\code{IntervalList$fromJSONString()}}
\item \href{#method-IntervalList-validateJSON}{\code{IntervalList$validateJSON()}}
\item \href{#method-IntervalList-toString}{\code{IntervalList$toString()}}
\item \href{#method-IntervalList-isValid}{\code{IntervalList$isValid()}}
\item \href{#method-IntervalList-getInvalidFields}{\code{IntervalList$getInvalidFields()}}
\item \href{#method-IntervalList-print}{\code{IntervalList$print()}}
\item \href{#method-IntervalList-clone}{\code{IntervalList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntervalList-new"></a>}}
\if{latex}{\out{\hypertarget{method-IntervalList-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new IntervalList class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalList$new(metadata = NULL, pagination = NULL, data = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{metadata}

\item{\code{pagination}}{pagination}

\item{\code{data}}{data}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntervalList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-IntervalList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntervalList-toList"></a>}}
\if{latex}{\out{\hypertarget{method-IntervalList-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalList$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
IntervalList as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of IntervalList (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntervalList-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-IntervalList-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert IntervalList to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalList$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntervalList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-IntervalList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of IntervalList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalList$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of IntervalList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntervalList-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-IntervalList-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalList$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
IntervalList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntervalList-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-IntervalList-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of IntervalList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalList$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of IntervalList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntervalList-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-IntervalList-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to IntervalList and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalList$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntervalList-toString"></a>}}
\if{latex}{\out{\hypertarget{method-IntervalList-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalList$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of IntervalList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntervalList-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-IntervalList-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalList$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntervalList-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-IntervalList-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalList$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntervalList-print"></a>}}
\if{latex}{\out{\hypertarget{method-IntervalList-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalList$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntervalList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IntervalList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
