% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs_api.R
\docType{class}
\name{JobsApi}
\alias{JobsApi}
\title{Jobs operations}
\format{
An \code{R6Class} generator object
}
\description{
JobsApi
}
\details{
Mednet EDC API

The Mednet EDC (Electronic Data Capture) REST API provides a single-point of access for reading data stored across iMednet data services. This specification documents all available endpoints in the latest version of the API. 

The version of the OpenAPI document: 1.0.15
Generated by: https://openapi-generator.tech
}
\examples{
\dontrun{
####################  GetJobStatus  ####################

library(openapi)
var_study_key <- "study_key_example" # character | Study key identifying the study context for the request
var_batch_id <- "batch_id_example" # character | Batch ID of the job to retrieve

#Retrieve job status by batch ID
api_instance <- JobsApi$new()

# Configure API key authorization: apiKeyAuth
api_instance$api_client$api_keys["x-api-key"] <- Sys.getenv("API_KEY")

# Configure API key authorization: securityKeyAuth
api_instance$api_client$api_keys["x-imn-security-key"] <- Sys.getenv("API_KEY")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$GetJobStatus(var_study_key, var_batch_iddata_file = "result.txt")
result <- api_instance$GetJobStatus(var_study_key, var_batch_id)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JobsApi-new}{\code{JobsApi$new()}}
\item \href{#method-JobsApi-GetJobStatus}{\code{JobsApi$GetJobStatus()}}
\item \href{#method-JobsApi-GetJobStatusWithHttpInfo}{\code{JobsApi$GetJobStatusWithHttpInfo()}}
\item \href{#method-JobsApi-clone}{\code{JobsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobsApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-JobsApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new JobsApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobsApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobsApi-GetJobStatus"></a>}}
\if{latex}{\out{\hypertarget{method-JobsApi-GetJobStatus}{}}}
\subsection{Method \code{GetJobStatus()}}{
Retrieve job status by batch ID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobsApi$GetJobStatus(study_key, batch_id, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study_key}}{Study key identifying the study context for the request}

\item{\code{batch_id}}{Batch ID of the job to retrieve}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Job
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobsApi-GetJobStatusWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-JobsApi-GetJobStatusWithHttpInfo}{}}}
\subsection{Method \code{GetJobStatusWithHttpInfo()}}{
Retrieve job status by batch ID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobsApi$GetJobStatusWithHttpInfo(study_key, batch_id, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study_key}}{Study key identifying the study context for the request}

\item{\code{batch_id}}{Batch ID of the job to retrieve}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (Job) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobsApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JobsApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
