% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\docType{class}
\name{Query}
\alias{Query}
\title{Query}
\format{
An \code{R6Class} generator object
}
\description{
Query Class

Query Class
}
\details{
Create a new Query
}
\examples{

## ------------------------------------------------
## Method `Query$toList`
## ------------------------------------------------

# convert array of Query (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{studyKey}}{Unique study key character [optional]}

\item{\code{subjectId}}{Mednet subject ID associated with the query (if applicable) integer [optional]}

\item{\code{subjectOid}}{Client-assigned subject OID (if applicable) character [optional]}

\item{\code{annotationType}}{Category/type of the query (e.g., subject, record, question) character [optional]}

\item{\code{annotationId}}{Unique system identifier for the query integer [optional]}

\item{\code{type}}{System text identifier for the query type/location (subject, record, question) character [optional]}

\item{\code{description}}{Description of the query (e.g., reason or context) character [optional]}

\item{\code{recordId}}{Record ID associated with the query (if applicable) integer [optional]}

\item{\code{variable}}{Variable name (field) associated with the query (if applicable) character [optional]}

\item{\code{subjectKey}}{Subject display ID associated with the query (if applicable) character [optional]}

\item{\code{dateCreated}}{Date when the query was created character [optional]}

\item{\code{dateModified}}{Date when the query was last modified character [optional]}

\item{\code{queryComments}}{History of comments/actions on the query list(\link{QueryComment}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Query-new}{\code{Query$new()}}
\item \href{#method-Query-toJSON}{\code{Query$toJSON()}}
\item \href{#method-Query-toList}{\code{Query$toList()}}
\item \href{#method-Query-toSimpleType}{\code{Query$toSimpleType()}}
\item \href{#method-Query-fromJSON}{\code{Query$fromJSON()}}
\item \href{#method-Query-toJSONString}{\code{Query$toJSONString()}}
\item \href{#method-Query-fromJSONString}{\code{Query$fromJSONString()}}
\item \href{#method-Query-validateJSON}{\code{Query$validateJSON()}}
\item \href{#method-Query-toString}{\code{Query$toString()}}
\item \href{#method-Query-isValid}{\code{Query$isValid()}}
\item \href{#method-Query-getInvalidFields}{\code{Query$getInvalidFields()}}
\item \href{#method-Query-print}{\code{Query$print()}}
\item \href{#method-Query-clone}{\code{Query$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-new"></a>}}
\if{latex}{\out{\hypertarget{method-Query-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Query class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$new(
  studyKey = NULL,
  subjectId = NULL,
  subjectOid = NULL,
  annotationType = NULL,
  annotationId = NULL,
  type = NULL,
  description = NULL,
  recordId = NULL,
  variable = NULL,
  subjectKey = NULL,
  dateCreated = NULL,
  dateModified = NULL,
  queryComments = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{studyKey}}{Unique study key}

\item{\code{subjectId}}{Mednet subject ID associated with the query (if applicable)}

\item{\code{subjectOid}}{Client-assigned subject OID (if applicable)}

\item{\code{annotationType}}{Category/type of the query (e.g., subject, record, question)}

\item{\code{annotationId}}{Unique system identifier for the query}

\item{\code{type}}{System text identifier for the query type/location (subject, record, question)}

\item{\code{description}}{Description of the query (e.g., reason or context)}

\item{\code{recordId}}{Record ID associated with the query (if applicable)}

\item{\code{variable}}{Variable name (field) associated with the query (if applicable)}

\item{\code{subjectKey}}{Subject display ID associated with the query (if applicable)}

\item{\code{dateCreated}}{Date when the query was created}

\item{\code{dateModified}}{Date when the query was last modified}

\item{\code{queryComments}}{History of comments/actions on the query}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Query-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Query-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Query as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of Query (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-Query-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert Query to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Query-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Query
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Query
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Query-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Query in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Query-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Query
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Query
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Query-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Query and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Query-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Query
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Query-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Query-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-print"></a>}}
\if{latex}{\out{\hypertarget{method-Query-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Query-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
