% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_list.R
\docType{class}
\name{QueryList}
\alias{QueryList}
\title{QueryList}
\format{
An \code{R6Class} generator object
}
\description{
QueryList Class

QueryList Class
}
\details{
Create a new QueryList
}
\examples{

## ------------------------------------------------
## Method `QueryList$toList`
## ------------------------------------------------

# convert array of QueryList (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{\link{Metadata} [optional]}

\item{\code{pagination}}{\link{Pagination} [optional]}

\item{\code{data}}{list(\link{Query}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QueryList-new}{\code{QueryList$new()}}
\item \href{#method-QueryList-toJSON}{\code{QueryList$toJSON()}}
\item \href{#method-QueryList-toList}{\code{QueryList$toList()}}
\item \href{#method-QueryList-toSimpleType}{\code{QueryList$toSimpleType()}}
\item \href{#method-QueryList-fromJSON}{\code{QueryList$fromJSON()}}
\item \href{#method-QueryList-toJSONString}{\code{QueryList$toJSONString()}}
\item \href{#method-QueryList-fromJSONString}{\code{QueryList$fromJSONString()}}
\item \href{#method-QueryList-validateJSON}{\code{QueryList$validateJSON()}}
\item \href{#method-QueryList-toString}{\code{QueryList$toString()}}
\item \href{#method-QueryList-isValid}{\code{QueryList$isValid()}}
\item \href{#method-QueryList-getInvalidFields}{\code{QueryList$getInvalidFields()}}
\item \href{#method-QueryList-print}{\code{QueryList$print()}}
\item \href{#method-QueryList-clone}{\code{QueryList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryList-new"></a>}}
\if{latex}{\out{\hypertarget{method-QueryList-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new QueryList class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryList$new(metadata = NULL, pagination = NULL, data = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{metadata}

\item{\code{pagination}}{pagination}

\item{\code{data}}{data}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QueryList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryList-toList"></a>}}
\if{latex}{\out{\hypertarget{method-QueryList-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryList$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
QueryList as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of QueryList (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryList-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-QueryList-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert QueryList to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryList$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QueryList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of QueryList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryList$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of QueryList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryList-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QueryList-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryList$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
QueryList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryList-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QueryList-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of QueryList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryList$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of QueryList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryList-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QueryList-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to QueryList and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryList$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryList-toString"></a>}}
\if{latex}{\out{\hypertarget{method-QueryList-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryList$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of QueryList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryList-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-QueryList-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryList$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryList-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-QueryList-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryList$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryList-print"></a>}}
\if{latex}{\out{\hypertarget{method-QueryList-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryList$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QueryList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
