% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_job_status.R
\docType{class}
\name{RecordJobStatus}
\alias{RecordJobStatus}
\title{RecordJobStatus}
\format{
An \code{R6Class} generator object
}
\description{
RecordJobStatus Class

RecordJobStatus Class
}
\details{
Create a new RecordJobStatus
}
\examples{

## ------------------------------------------------
## Method `RecordJobStatus$toList`
## ------------------------------------------------

# convert array of RecordJobStatus (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{jobId}}{Unique job identifier character [optional]}

\item{\code{batchId}}{Batch ID used to track the job (identical to jobId in most cases) character [optional]}

\item{\code{state}}{Current state of the job (e.g., created, completed) character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RecordJobStatus-new}{\code{RecordJobStatus$new()}}
\item \href{#method-RecordJobStatus-toJSON}{\code{RecordJobStatus$toJSON()}}
\item \href{#method-RecordJobStatus-toList}{\code{RecordJobStatus$toList()}}
\item \href{#method-RecordJobStatus-toSimpleType}{\code{RecordJobStatus$toSimpleType()}}
\item \href{#method-RecordJobStatus-fromJSON}{\code{RecordJobStatus$fromJSON()}}
\item \href{#method-RecordJobStatus-toJSONString}{\code{RecordJobStatus$toJSONString()}}
\item \href{#method-RecordJobStatus-fromJSONString}{\code{RecordJobStatus$fromJSONString()}}
\item \href{#method-RecordJobStatus-validateJSON}{\code{RecordJobStatus$validateJSON()}}
\item \href{#method-RecordJobStatus-toString}{\code{RecordJobStatus$toString()}}
\item \href{#method-RecordJobStatus-isValid}{\code{RecordJobStatus$isValid()}}
\item \href{#method-RecordJobStatus-getInvalidFields}{\code{RecordJobStatus$getInvalidFields()}}
\item \href{#method-RecordJobStatus-print}{\code{RecordJobStatus$print()}}
\item \href{#method-RecordJobStatus-clone}{\code{RecordJobStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordJobStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-RecordJobStatus-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new RecordJobStatus class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordJobStatus$new(jobId = NULL, batchId = NULL, state = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{jobId}}{Unique job identifier}

\item{\code{batchId}}{Batch ID used to track the job (identical to jobId in most cases)}

\item{\code{state}}{Current state of the job (e.g., created, completed)}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordJobStatus-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RecordJobStatus-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordJobStatus$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordJobStatus-toList"></a>}}
\if{latex}{\out{\hypertarget{method-RecordJobStatus-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordJobStatus$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
RecordJobStatus as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of RecordJobStatus (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordJobStatus-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-RecordJobStatus-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert RecordJobStatus to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordJobStatus$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordJobStatus-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RecordJobStatus-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of RecordJobStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordJobStatus$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RecordJobStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordJobStatus-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RecordJobStatus-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordJobStatus$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
RecordJobStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordJobStatus-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RecordJobStatus-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of RecordJobStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordJobStatus$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RecordJobStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordJobStatus-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RecordJobStatus-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to RecordJobStatus and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordJobStatus$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordJobStatus-toString"></a>}}
\if{latex}{\out{\hypertarget{method-RecordJobStatus-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordJobStatus$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of RecordJobStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordJobStatus-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-RecordJobStatus-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordJobStatus$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordJobStatus-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-RecordJobStatus-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordJobStatus$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordJobStatus-print"></a>}}
\if{latex}{\out{\hypertarget{method-RecordJobStatus-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordJobStatus$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordJobStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RecordJobStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordJobStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
