% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_list.R
\docType{class}
\name{RecordList}
\alias{RecordList}
\title{RecordList}
\format{
An \code{R6Class} generator object
}
\description{
RecordList Class

RecordList Class
}
\details{
Create a new RecordList
}
\examples{

## ------------------------------------------------
## Method `RecordList$toList`
## ------------------------------------------------

# convert array of RecordList (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{\link{Metadata} [optional]}

\item{\code{pagination}}{\link{Pagination} [optional]}

\item{\code{data}}{list(\link{Record}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RecordList-new}{\code{RecordList$new()}}
\item \href{#method-RecordList-toJSON}{\code{RecordList$toJSON()}}
\item \href{#method-RecordList-toList}{\code{RecordList$toList()}}
\item \href{#method-RecordList-toSimpleType}{\code{RecordList$toSimpleType()}}
\item \href{#method-RecordList-fromJSON}{\code{RecordList$fromJSON()}}
\item \href{#method-RecordList-toJSONString}{\code{RecordList$toJSONString()}}
\item \href{#method-RecordList-fromJSONString}{\code{RecordList$fromJSONString()}}
\item \href{#method-RecordList-validateJSON}{\code{RecordList$validateJSON()}}
\item \href{#method-RecordList-toString}{\code{RecordList$toString()}}
\item \href{#method-RecordList-isValid}{\code{RecordList$isValid()}}
\item \href{#method-RecordList-getInvalidFields}{\code{RecordList$getInvalidFields()}}
\item \href{#method-RecordList-print}{\code{RecordList$print()}}
\item \href{#method-RecordList-clone}{\code{RecordList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordList-new"></a>}}
\if{latex}{\out{\hypertarget{method-RecordList-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new RecordList class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordList$new(metadata = NULL, pagination = NULL, data = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{metadata}

\item{\code{pagination}}{pagination}

\item{\code{data}}{data}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RecordList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordList-toList"></a>}}
\if{latex}{\out{\hypertarget{method-RecordList-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordList$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
RecordList as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of RecordList (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordList-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-RecordList-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert RecordList to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordList$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RecordList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of RecordList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordList$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RecordList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordList-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RecordList-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordList$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
RecordList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordList-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RecordList-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of RecordList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordList$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RecordList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordList-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RecordList-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to RecordList and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordList$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordList-toString"></a>}}
\if{latex}{\out{\hypertarget{method-RecordList-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordList$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of RecordList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordList-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-RecordList-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordList$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordList-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-RecordList-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordList$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordList-print"></a>}}
\if{latex}{\out{\hypertarget{method-RecordList-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordList$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RecordList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
