% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_revision_list.R
\docType{class}
\name{RecordRevisionList}
\alias{RecordRevisionList}
\title{RecordRevisionList}
\format{
An \code{R6Class} generator object
}
\description{
RecordRevisionList Class

RecordRevisionList Class
}
\details{
Create a new RecordRevisionList
}
\examples{

## ------------------------------------------------
## Method `RecordRevisionList$toList`
## ------------------------------------------------

# convert array of RecordRevisionList (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{\link{Metadata} [optional]}

\item{\code{pagination}}{\link{Pagination} [optional]}

\item{\code{data}}{list(\link{RecordRevision}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RecordRevisionList-new}{\code{RecordRevisionList$new()}}
\item \href{#method-RecordRevisionList-toJSON}{\code{RecordRevisionList$toJSON()}}
\item \href{#method-RecordRevisionList-toList}{\code{RecordRevisionList$toList()}}
\item \href{#method-RecordRevisionList-toSimpleType}{\code{RecordRevisionList$toSimpleType()}}
\item \href{#method-RecordRevisionList-fromJSON}{\code{RecordRevisionList$fromJSON()}}
\item \href{#method-RecordRevisionList-toJSONString}{\code{RecordRevisionList$toJSONString()}}
\item \href{#method-RecordRevisionList-fromJSONString}{\code{RecordRevisionList$fromJSONString()}}
\item \href{#method-RecordRevisionList-validateJSON}{\code{RecordRevisionList$validateJSON()}}
\item \href{#method-RecordRevisionList-toString}{\code{RecordRevisionList$toString()}}
\item \href{#method-RecordRevisionList-isValid}{\code{RecordRevisionList$isValid()}}
\item \href{#method-RecordRevisionList-getInvalidFields}{\code{RecordRevisionList$getInvalidFields()}}
\item \href{#method-RecordRevisionList-print}{\code{RecordRevisionList$print()}}
\item \href{#method-RecordRevisionList-clone}{\code{RecordRevisionList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordRevisionList-new"></a>}}
\if{latex}{\out{\hypertarget{method-RecordRevisionList-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new RecordRevisionList class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordRevisionList$new(metadata = NULL, pagination = NULL, data = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{metadata}

\item{\code{pagination}}{pagination}

\item{\code{data}}{data}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordRevisionList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RecordRevisionList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordRevisionList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordRevisionList-toList"></a>}}
\if{latex}{\out{\hypertarget{method-RecordRevisionList-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordRevisionList$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
RecordRevisionList as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of RecordRevisionList (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordRevisionList-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-RecordRevisionList-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert RecordRevisionList to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordRevisionList$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordRevisionList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RecordRevisionList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of RecordRevisionList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordRevisionList$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RecordRevisionList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordRevisionList-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RecordRevisionList-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordRevisionList$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
RecordRevisionList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordRevisionList-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RecordRevisionList-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of RecordRevisionList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordRevisionList$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RecordRevisionList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordRevisionList-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RecordRevisionList-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to RecordRevisionList and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordRevisionList$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordRevisionList-toString"></a>}}
\if{latex}{\out{\hypertarget{method-RecordRevisionList-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordRevisionList$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of RecordRevisionList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordRevisionList-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-RecordRevisionList-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordRevisionList$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordRevisionList-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-RecordRevisionList-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordRevisionList$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordRevisionList-print"></a>}}
\if{latex}{\out{\hypertarget{method-RecordRevisionList-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordRevisionList$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordRevisionList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RecordRevisionList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordRevisionList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
