% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records_api.R
\docType{class}
\name{RecordsApi}
\alias{RecordsApi}
\title{Records operations}
\format{
An \code{R6Class} generator object
}
\description{
RecordsApi
}
\details{
Mednet EDC API

The Mednet EDC (Electronic Data Capture) REST API provides a single-point of access for reading data stored across iMednet data services. This specification documents all available endpoints in the latest version of the API. 

The version of the OpenAPI document: 1.0.15
Generated by: https://openapi-generator.tech
}
\examples{
\dontrun{
####################  CreateRecords  ####################

library(openapi)
var_study_key <- "study_key_example" # character | Study key identifying the study context for the request
var_components_schemas_record_create_request_item <- c(components_schemas_RecordCreateRequest_item$new("formKey_example", c(key = TODO), 123, "siteName_example", 123, "subjectKey_example", 123, "subjectOid_example", "intervalName_example", 123, 123, "recordOid_example")) # array[ComponentsSchemasRecordCreateRequestItem] | An array of record objects defining the operation to perform (subject registration, record creation, or scheduled record update).

#Add new record or update subject/record data
api_instance <- RecordsApi$new()

# Configure API key authorization: apiKeyAuth
api_instance$api_client$api_keys["x-api-key"] <- Sys.getenv("API_KEY")

# Configure API key authorization: securityKeyAuth
api_instance$api_client$api_keys["x-imn-security-key"] <- Sys.getenv("API_KEY")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$CreateRecords(var_study_key, var_components_schemas_record_create_request_itemdata_file = "result.txt")
result <- api_instance$CreateRecords(var_study_key, var_components_schemas_record_create_request_item)
dput(result)


####################  ListRecords  ####################

library(openapi)
var_study_key <- "study_key_example" # character | Study key identifying the study context for the request
var_page <- 0 # integer | Page index to retrieve (0-based) (Optional)
var_size <- 25 # integer | Number of items to return per page (max 500) (Optional)
var_sort <- "sort_example" # character | Sorting criteria in the format `property,ASC` or `property,DESC`. Can be repeated. (Optional)
var_filter <- "filter_example" # character | Optional filter criteria to apply, in the format `attribute==value` with support for >, >=, <, <=, !=. Multiple criteria can be combined with `;` (AND) or `,` (OR). (Optional)
var_record_data_filter <- "record_data_filter_example" # character | Optional filter criteria to apply on recordData fields (question responses). Supports ==, !=, <, <=, >, >=, =~ (contains) and can combine multiple criteria with `;` (AND) or `,` (OR). (Optional)

#List records (eCRF instances) in a study
api_instance <- RecordsApi$new()

# Configure API key authorization: apiKeyAuth
api_instance$api_client$api_keys["x-api-key"] <- Sys.getenv("API_KEY")

# Configure API key authorization: securityKeyAuth
api_instance$api_client$api_keys["x-imn-security-key"] <- Sys.getenv("API_KEY")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$ListRecords(var_study_key, page = var_page, size = var_size, sort = var_sort, filter = var_filter, record_data_filter = var_record_data_filterdata_file = "result.txt")
result <- api_instance$ListRecords(var_study_key, page = var_page, size = var_size, sort = var_sort, filter = var_filter, record_data_filter = var_record_data_filter)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RecordsApi-new}{\code{RecordsApi$new()}}
\item \href{#method-RecordsApi-CreateRecords}{\code{RecordsApi$CreateRecords()}}
\item \href{#method-RecordsApi-CreateRecordsWithHttpInfo}{\code{RecordsApi$CreateRecordsWithHttpInfo()}}
\item \href{#method-RecordsApi-ListRecords}{\code{RecordsApi$ListRecords()}}
\item \href{#method-RecordsApi-ListRecordsWithHttpInfo}{\code{RecordsApi$ListRecordsWithHttpInfo()}}
\item \href{#method-RecordsApi-clone}{\code{RecordsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordsApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-RecordsApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new RecordsApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordsApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordsApi-CreateRecords"></a>}}
\if{latex}{\out{\hypertarget{method-RecordsApi-CreateRecords}{}}}
\subsection{Method \code{CreateRecords()}}{
Add new record or update subject/record data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordsApi$CreateRecords(
  study_key,
  components_schemas_record_create_request_item,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study_key}}{Study key identifying the study context for the request}

\item{\code{components_schemas_record_create_request_item}}{An array of record objects defining the operation to perform (subject registration, record creation, or scheduled record update).}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
RecordJobStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordsApi-CreateRecordsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-RecordsApi-CreateRecordsWithHttpInfo}{}}}
\subsection{Method \code{CreateRecordsWithHttpInfo()}}{
Add new record or update subject/record data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordsApi$CreateRecordsWithHttpInfo(
  study_key,
  components_schemas_record_create_request_item,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study_key}}{Study key identifying the study context for the request}

\item{\code{components_schemas_record_create_request_item}}{An array of record objects defining the operation to perform (subject registration, record creation, or scheduled record update).}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (RecordJobStatus) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordsApi-ListRecords"></a>}}
\if{latex}{\out{\hypertarget{method-RecordsApi-ListRecords}{}}}
\subsection{Method \code{ListRecords()}}{
List records (eCRF instances) in a study
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordsApi$ListRecords(
  study_key,
  page = 0,
  size = 25,
  sort = NULL,
  filter = NULL,
  record_data_filter = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study_key}}{Study key identifying the study context for the request}

\item{\code{page}}{(optional) Page index to retrieve (0-based) (default value: 0)}

\item{\code{size}}{(optional) Number of items to return per page (max 500) (default value: 25)}

\item{\code{sort}}{(optional) Sorting criteria in the format `property,ASC` or `property,DESC`. Can be repeated.}

\item{\code{filter}}{(optional) Optional filter criteria to apply, in the format `attribute==value` with support for >, >=, <, <=, !=. Multiple criteria can be combined with `;` (AND) or `,` (OR).}

\item{\code{record_data_filter}}{(optional) Optional filter criteria to apply on recordData fields (question responses). Supports ==, !=, <, <=, >, >=, =~ (contains) and can combine multiple criteria with `;` (AND) or `,` (OR).}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
RecordList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordsApi-ListRecordsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-RecordsApi-ListRecordsWithHttpInfo}{}}}
\subsection{Method \code{ListRecordsWithHttpInfo()}}{
List records (eCRF instances) in a study
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordsApi$ListRecordsWithHttpInfo(
  study_key,
  page = 0,
  size = 25,
  sort = NULL,
  filter = NULL,
  record_data_filter = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study_key}}{Study key identifying the study context for the request}

\item{\code{page}}{(optional) Page index to retrieve (0-based) (default value: 0)}

\item{\code{size}}{(optional) Number of items to return per page (max 500) (default value: 25)}

\item{\code{sort}}{(optional) Sorting criteria in the format `property,ASC` or `property,DESC`. Can be repeated.}

\item{\code{filter}}{(optional) Optional filter criteria to apply, in the format `attribute==value` with support for >, >=, <, <=, !=. Multiple criteria can be combined with `;` (AND) or `,` (OR).}

\item{\code{record_data_filter}}{(optional) Optional filter criteria to apply on recordData fields (question responses). Supports ==, !=, <, <=, >, >=, =~ (contains) and can combine multiple criteria with `;` (AND) or `,` (OR).}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (RecordList) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordsApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RecordsApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
