% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_list.R
\docType{class}
\name{SiteList}
\alias{SiteList}
\title{SiteList}
\format{
An \code{R6Class} generator object
}
\description{
SiteList Class

SiteList Class
}
\details{
Create a new SiteList
}
\examples{

## ------------------------------------------------
## Method `SiteList$toList`
## ------------------------------------------------

# convert array of SiteList (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{\link{Metadata} [optional]}

\item{\code{pagination}}{\link{Pagination} [optional]}

\item{\code{data}}{list(\link{Site}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SiteList-new}{\code{SiteList$new()}}
\item \href{#method-SiteList-toJSON}{\code{SiteList$toJSON()}}
\item \href{#method-SiteList-toList}{\code{SiteList$toList()}}
\item \href{#method-SiteList-toSimpleType}{\code{SiteList$toSimpleType()}}
\item \href{#method-SiteList-fromJSON}{\code{SiteList$fromJSON()}}
\item \href{#method-SiteList-toJSONString}{\code{SiteList$toJSONString()}}
\item \href{#method-SiteList-fromJSONString}{\code{SiteList$fromJSONString()}}
\item \href{#method-SiteList-validateJSON}{\code{SiteList$validateJSON()}}
\item \href{#method-SiteList-toString}{\code{SiteList$toString()}}
\item \href{#method-SiteList-isValid}{\code{SiteList$isValid()}}
\item \href{#method-SiteList-getInvalidFields}{\code{SiteList$getInvalidFields()}}
\item \href{#method-SiteList-print}{\code{SiteList$print()}}
\item \href{#method-SiteList-clone}{\code{SiteList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SiteList-new"></a>}}
\if{latex}{\out{\hypertarget{method-SiteList-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new SiteList class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SiteList$new(metadata = NULL, pagination = NULL, data = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{metadata}

\item{\code{pagination}}{pagination}

\item{\code{data}}{data}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SiteList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SiteList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SiteList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SiteList-toList"></a>}}
\if{latex}{\out{\hypertarget{method-SiteList-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SiteList$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
SiteList as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of SiteList (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SiteList-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-SiteList-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert SiteList to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SiteList$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SiteList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SiteList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of SiteList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SiteList$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SiteList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SiteList-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SiteList-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SiteList$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
SiteList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SiteList-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SiteList-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of SiteList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SiteList$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SiteList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SiteList-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SiteList-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to SiteList and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SiteList$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SiteList-toString"></a>}}
\if{latex}{\out{\hypertarget{method-SiteList-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SiteList$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of SiteList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SiteList-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-SiteList-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SiteList$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SiteList-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-SiteList-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SiteList$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SiteList-print"></a>}}
\if{latex}{\out{\hypertarget{method-SiteList-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SiteList$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SiteList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SiteList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SiteList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
