% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites_api.R
\docType{class}
\name{SitesApi}
\alias{SitesApi}
\title{Sites operations}
\format{
An \code{R6Class} generator object
}
\description{
SitesApi
}
\details{
Mednet EDC API

The Mednet EDC (Electronic Data Capture) REST API provides a single-point of access for reading data stored across iMednet data services. This specification documents all available endpoints in the latest version of the API. 

The version of the OpenAPI document: 1.0.15
Generated by: https://openapi-generator.tech
}
\examples{
\dontrun{
####################  ListSites  ####################

library(openapi)
var_study_key <- "study_key_example" # character | Study key identifying the study context for the request
var_page <- 0 # integer | Page index to retrieve (0-based) (Optional)
var_size <- 25 # integer | Number of items to return per page (max 500) (Optional)
var_sort <- "sort_example" # character | Sorting criteria in the format `property,ASC` or `property,DESC`. Can be repeated. (Optional)
var_filter <- "filter_example" # character | Optional filter criteria to apply, in the format `attribute==value` with support for >, >=, <, <=, !=. Multiple criteria can be combined with `;` (AND) or `,` (OR). (Optional)

#List sites for a study
api_instance <- SitesApi$new()

# Configure API key authorization: apiKeyAuth
api_instance$api_client$api_keys["x-api-key"] <- Sys.getenv("API_KEY")

# Configure API key authorization: securityKeyAuth
api_instance$api_client$api_keys["x-imn-security-key"] <- Sys.getenv("API_KEY")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$ListSites(var_study_key, page = var_page, size = var_size, sort = var_sort, filter = var_filterdata_file = "result.txt")
result <- api_instance$ListSites(var_study_key, page = var_page, size = var_size, sort = var_sort, filter = var_filter)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SitesApi-new}{\code{SitesApi$new()}}
\item \href{#method-SitesApi-ListSites}{\code{SitesApi$ListSites()}}
\item \href{#method-SitesApi-ListSitesWithHttpInfo}{\code{SitesApi$ListSitesWithHttpInfo()}}
\item \href{#method-SitesApi-clone}{\code{SitesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SitesApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-SitesApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new SitesApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SitesApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SitesApi-ListSites"></a>}}
\if{latex}{\out{\hypertarget{method-SitesApi-ListSites}{}}}
\subsection{Method \code{ListSites()}}{
List sites for a study
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SitesApi$ListSites(
  study_key,
  page = 0,
  size = 25,
  sort = NULL,
  filter = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study_key}}{Study key identifying the study context for the request}

\item{\code{page}}{(optional) Page index to retrieve (0-based) (default value: 0)}

\item{\code{size}}{(optional) Number of items to return per page (max 500) (default value: 25)}

\item{\code{sort}}{(optional) Sorting criteria in the format `property,ASC` or `property,DESC`. Can be repeated.}

\item{\code{filter}}{(optional) Optional filter criteria to apply, in the format `attribute==value` with support for >, >=, <, <=, !=. Multiple criteria can be combined with `;` (AND) or `,` (OR).}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
SiteList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SitesApi-ListSitesWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SitesApi-ListSitesWithHttpInfo}{}}}
\subsection{Method \code{ListSitesWithHttpInfo()}}{
List sites for a study
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SitesApi$ListSitesWithHttpInfo(
  study_key,
  page = 0,
  size = 25,
  sort = NULL,
  filter = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study_key}}{Study key identifying the study context for the request}

\item{\code{page}}{(optional) Page index to retrieve (0-based) (default value: 0)}

\item{\code{size}}{(optional) Number of items to return per page (max 500) (default value: 25)}

\item{\code{sort}}{(optional) Sorting criteria in the format `property,ASC` or `property,DESC`. Can be repeated.}

\item{\code{filter}}{(optional) Optional filter criteria to apply, in the format `attribute==value` with support for >, >=, <, <=, !=. Multiple criteria can be combined with `;` (AND) or `,` (OR).}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (SiteList) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SitesApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SitesApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SitesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
