% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\docType{class}
\name{Sort}
\alias{Sort}
\title{Sort}
\format{
An \code{R6Class} generator object
}
\description{
Sort Class

Sort Class
}
\details{
Create a new Sort
}
\examples{

## ------------------------------------------------
## Method `Sort$toList`
## ------------------------------------------------

# convert array of Sort (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{property}}{Name of the property by which the result is sorted character [optional]}

\item{\code{direction}}{Sort direction (ASC or DESC) character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Sort-new}{\code{Sort$new()}}
\item \href{#method-Sort-toJSON}{\code{Sort$toJSON()}}
\item \href{#method-Sort-toList}{\code{Sort$toList()}}
\item \href{#method-Sort-toSimpleType}{\code{Sort$toSimpleType()}}
\item \href{#method-Sort-fromJSON}{\code{Sort$fromJSON()}}
\item \href{#method-Sort-toJSONString}{\code{Sort$toJSONString()}}
\item \href{#method-Sort-fromJSONString}{\code{Sort$fromJSONString()}}
\item \href{#method-Sort-validateJSON}{\code{Sort$validateJSON()}}
\item \href{#method-Sort-toString}{\code{Sort$toString()}}
\item \href{#method-Sort-isValid}{\code{Sort$isValid()}}
\item \href{#method-Sort-getInvalidFields}{\code{Sort$getInvalidFields()}}
\item \href{#method-Sort-print}{\code{Sort$print()}}
\item \href{#method-Sort-clone}{\code{Sort$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sort-new"></a>}}
\if{latex}{\out{\hypertarget{method-Sort-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Sort class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sort$new(property = NULL, direction = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{property}}{Name of the property by which the result is sorted}

\item{\code{direction}}{Sort direction (ASC or DESC)}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sort-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Sort-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sort$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sort-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Sort-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sort$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Sort as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of Sort (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sort-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-Sort-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert Sort to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sort$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sort-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Sort-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Sort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sort$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Sort
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sort-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Sort-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sort$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Sort in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sort-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Sort-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Sort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sort$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Sort
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sort-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Sort-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Sort and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sort$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sort-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Sort-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sort$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Sort
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sort-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Sort-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sort$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sort-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Sort-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sort$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sort-print"></a>}}
\if{latex}{\out{\hypertarget{method-Sort-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sort$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sort-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Sort-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sort$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
