% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies_api.R
\docType{class}
\name{StudiesApi}
\alias{StudiesApi}
\title{Studies operations}
\format{
An \code{R6Class} generator object
}
\description{
StudiesApi
}
\details{
Mednet EDC API

The Mednet EDC (Electronic Data Capture) REST API provides a single-point of access for reading data stored across iMednet data services. This specification documents all available endpoints in the latest version of the API. 

The version of the OpenAPI document: 1.0.15
Generated by: https://openapi-generator.tech
}
\examples{
\dontrun{
####################  ListStudies  ####################

library(openapi)
var_page <- 0 # integer | Page index to retrieve (0-based) (Optional)
var_size <- 25 # integer | Number of items to return per page (max 500) (Optional)
var_sort <- "sort_example" # character | Sorting criteria in the format `property,ASC` or `property,DESC`. Can be repeated. (Optional)
var_filter <- "filter_example" # character | Optional filter criteria to apply, in the format `attribute==value` with support for >, >=, <, <=, !=. Multiple criteria can be combined with `;` (AND) or `,` (OR). (Optional)

#List studies accessible by API key
api_instance <- StudiesApi$new()

# Configure API key authorization: apiKeyAuth
api_instance$api_client$api_keys["x-api-key"] <- Sys.getenv("API_KEY")

# Configure API key authorization: securityKeyAuth
api_instance$api_client$api_keys["x-imn-security-key"] <- Sys.getenv("API_KEY")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$ListStudies(page = var_page, size = var_size, sort = var_sort, filter = var_filterdata_file = "result.txt")
result <- api_instance$ListStudies(page = var_page, size = var_size, sort = var_sort, filter = var_filter)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StudiesApi-new}{\code{StudiesApi$new()}}
\item \href{#method-StudiesApi-ListStudies}{\code{StudiesApi$ListStudies()}}
\item \href{#method-StudiesApi-ListStudiesWithHttpInfo}{\code{StudiesApi$ListStudiesWithHttpInfo()}}
\item \href{#method-StudiesApi-clone}{\code{StudiesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudiesApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-StudiesApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new StudiesApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudiesApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudiesApi-ListStudies"></a>}}
\if{latex}{\out{\hypertarget{method-StudiesApi-ListStudies}{}}}
\subsection{Method \code{ListStudies()}}{
List studies accessible by API key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudiesApi$ListStudies(
  page = 0,
  size = 25,
  sort = NULL,
  filter = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{page}}{(optional) Page index to retrieve (0-based) (default value: 0)}

\item{\code{size}}{(optional) Number of items to return per page (max 500) (default value: 25)}

\item{\code{sort}}{(optional) Sorting criteria in the format `property,ASC` or `property,DESC`. Can be repeated.}

\item{\code{filter}}{(optional) Optional filter criteria to apply, in the format `attribute==value` with support for >, >=, <, <=, !=. Multiple criteria can be combined with `;` (AND) or `,` (OR).}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
StudyList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudiesApi-ListStudiesWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-StudiesApi-ListStudiesWithHttpInfo}{}}}
\subsection{Method \code{ListStudiesWithHttpInfo()}}{
List studies accessible by API key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudiesApi$ListStudiesWithHttpInfo(
  page = 0,
  size = 25,
  sort = NULL,
  filter = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{page}}{(optional) Page index to retrieve (0-based) (default value: 0)}

\item{\code{size}}{(optional) Number of items to return per page (max 500) (default value: 25)}

\item{\code{sort}}{(optional) Sorting criteria in the format `property,ASC` or `property,DESC`. Can be repeated.}

\item{\code{filter}}{(optional) Optional filter criteria to apply, in the format `attribute==value` with support for >, >=, <, <=, !=. Multiple criteria can be combined with `;` (AND) or `,` (OR).}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (StudyList) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudiesApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StudiesApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudiesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
