% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\docType{class}
\name{Study}
\alias{Study}
\title{Study}
\format{
An \code{R6Class} generator object
}
\description{
Study Class

Study Class
}
\details{
Create a new Study
}
\examples{

## ------------------------------------------------
## Method `Study$toList`
## ------------------------------------------------

# convert array of Study (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sponsorKey}}{Sponsor key that this study belongs to character [optional]}

\item{\code{studyKey}}{Unique study key character [optional]}

\item{\code{studyId}}{Mednet study ID (internal numeric identifier) integer [optional]}

\item{\code{studyName}}{Name of the study character [optional]}

\item{\code{studyDescription}}{Description of the study character [optional]}

\item{\code{studyType}}{Type of study (e.g., STUDY) character [optional]}

\item{\code{dateCreated}}{Date when the study record was created character [optional]}

\item{\code{dateModified}}{Date when the study record was last modified character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Study-new}{\code{Study$new()}}
\item \href{#method-Study-toJSON}{\code{Study$toJSON()}}
\item \href{#method-Study-toList}{\code{Study$toList()}}
\item \href{#method-Study-toSimpleType}{\code{Study$toSimpleType()}}
\item \href{#method-Study-fromJSON}{\code{Study$fromJSON()}}
\item \href{#method-Study-toJSONString}{\code{Study$toJSONString()}}
\item \href{#method-Study-fromJSONString}{\code{Study$fromJSONString()}}
\item \href{#method-Study-validateJSON}{\code{Study$validateJSON()}}
\item \href{#method-Study-toString}{\code{Study$toString()}}
\item \href{#method-Study-isValid}{\code{Study$isValid()}}
\item \href{#method-Study-getInvalidFields}{\code{Study$getInvalidFields()}}
\item \href{#method-Study-print}{\code{Study$print()}}
\item \href{#method-Study-clone}{\code{Study$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-new"></a>}}
\if{latex}{\out{\hypertarget{method-Study-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Study class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$new(
  sponsorKey = NULL,
  studyKey = NULL,
  studyId = NULL,
  studyName = NULL,
  studyDescription = NULL,
  studyType = NULL,
  dateCreated = NULL,
  dateModified = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sponsorKey}}{Sponsor key that this study belongs to}

\item{\code{studyKey}}{Unique study key}

\item{\code{studyId}}{Mednet study ID (internal numeric identifier)}

\item{\code{studyName}}{Name of the study}

\item{\code{studyDescription}}{Description of the study}

\item{\code{studyType}}{Type of study (e.g., STUDY)}

\item{\code{dateCreated}}{Date when the study record was created}

\item{\code{dateModified}}{Date when the study record was last modified}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Study-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Study-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Study as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of Study (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-Study-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert Study to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Study-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Study
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Study
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Study-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Study in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Study-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Study
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Study
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Study-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Study and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Study-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Study
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Study-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Study-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-print"></a>}}
\if{latex}{\out{\hypertarget{method-Study-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Study-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
