% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_list.R
\docType{class}
\name{StudyList}
\alias{StudyList}
\title{StudyList}
\format{
An \code{R6Class} generator object
}
\description{
StudyList Class

StudyList Class
}
\details{
Create a new StudyList
}
\examples{

## ------------------------------------------------
## Method `StudyList$toList`
## ------------------------------------------------

# convert array of StudyList (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{\link{Metadata} [optional]}

\item{\code{pagination}}{\link{Pagination} [optional]}

\item{\code{data}}{list(\link{Study}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StudyList-new}{\code{StudyList$new()}}
\item \href{#method-StudyList-toJSON}{\code{StudyList$toJSON()}}
\item \href{#method-StudyList-toList}{\code{StudyList$toList()}}
\item \href{#method-StudyList-toSimpleType}{\code{StudyList$toSimpleType()}}
\item \href{#method-StudyList-fromJSON}{\code{StudyList$fromJSON()}}
\item \href{#method-StudyList-toJSONString}{\code{StudyList$toJSONString()}}
\item \href{#method-StudyList-fromJSONString}{\code{StudyList$fromJSONString()}}
\item \href{#method-StudyList-validateJSON}{\code{StudyList$validateJSON()}}
\item \href{#method-StudyList-toString}{\code{StudyList$toString()}}
\item \href{#method-StudyList-isValid}{\code{StudyList$isValid()}}
\item \href{#method-StudyList-getInvalidFields}{\code{StudyList$getInvalidFields()}}
\item \href{#method-StudyList-print}{\code{StudyList$print()}}
\item \href{#method-StudyList-clone}{\code{StudyList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyList-new"></a>}}
\if{latex}{\out{\hypertarget{method-StudyList-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new StudyList class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyList$new(metadata = NULL, pagination = NULL, data = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{metadata}

\item{\code{pagination}}{pagination}

\item{\code{data}}{data}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StudyList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyList-toList"></a>}}
\if{latex}{\out{\hypertarget{method-StudyList-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyList$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
StudyList as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of StudyList (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyList-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-StudyList-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert StudyList to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyList$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StudyList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of StudyList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyList$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of StudyList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyList-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StudyList-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyList$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
StudyList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyList-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StudyList-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of StudyList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyList$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of StudyList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyList-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StudyList-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to StudyList and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyList$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyList-toString"></a>}}
\if{latex}{\out{\hypertarget{method-StudyList-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyList$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of StudyList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyList-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-StudyList-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyList$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyList-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-StudyList-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyList$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyList-print"></a>}}
\if{latex}{\out{\hypertarget{method-StudyList-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyList$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StudyList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
