% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_list.R
\docType{class}
\name{UserList}
\alias{UserList}
\title{UserList}
\format{
An \code{R6Class} generator object
}
\description{
UserList Class

UserList Class
}
\details{
Create a new UserList
}
\examples{

## ------------------------------------------------
## Method `UserList$toList`
## ------------------------------------------------

# convert array of UserList (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{\link{Metadata} [optional]}

\item{\code{pagination}}{\link{Pagination} [optional]}

\item{\code{data}}{list(\link{User}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserList-new}{\code{UserList$new()}}
\item \href{#method-UserList-toJSON}{\code{UserList$toJSON()}}
\item \href{#method-UserList-toList}{\code{UserList$toList()}}
\item \href{#method-UserList-toSimpleType}{\code{UserList$toSimpleType()}}
\item \href{#method-UserList-fromJSON}{\code{UserList$fromJSON()}}
\item \href{#method-UserList-toJSONString}{\code{UserList$toJSONString()}}
\item \href{#method-UserList-fromJSONString}{\code{UserList$fromJSONString()}}
\item \href{#method-UserList-validateJSON}{\code{UserList$validateJSON()}}
\item \href{#method-UserList-toString}{\code{UserList$toString()}}
\item \href{#method-UserList-isValid}{\code{UserList$isValid()}}
\item \href{#method-UserList-getInvalidFields}{\code{UserList$getInvalidFields()}}
\item \href{#method-UserList-print}{\code{UserList$print()}}
\item \href{#method-UserList-clone}{\code{UserList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserList-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserList-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UserList class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserList$new(metadata = NULL, pagination = NULL, data = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{metadata}

\item{\code{pagination}}{pagination}

\item{\code{data}}{data}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserList-toList"></a>}}
\if{latex}{\out{\hypertarget{method-UserList-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserList$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserList as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of UserList (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserList-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-UserList-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert UserList to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserList$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UserList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserList$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserList-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserList-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserList$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
UserList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserList-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserList-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UserList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserList$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserList-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserList-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UserList and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserList$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserList-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UserList-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserList$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UserList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserList-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UserList-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserList$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserList-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UserList-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserList$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserList-print"></a>}}
\if{latex}{\out{\hypertarget{method-UserList-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserList$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
