% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_list.R
\docType{class}
\name{VariableList}
\alias{VariableList}
\title{VariableList}
\format{
An \code{R6Class} generator object
}
\description{
VariableList Class

VariableList Class
}
\details{
Create a new VariableList
}
\examples{

## ------------------------------------------------
## Method `VariableList$toList`
## ------------------------------------------------

# convert array of VariableList (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{\link{Metadata} [optional]}

\item{\code{pagination}}{\link{Pagination} [optional]}

\item{\code{data}}{list(\link{Variable}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VariableList-new}{\code{VariableList$new()}}
\item \href{#method-VariableList-toJSON}{\code{VariableList$toJSON()}}
\item \href{#method-VariableList-toList}{\code{VariableList$toList()}}
\item \href{#method-VariableList-toSimpleType}{\code{VariableList$toSimpleType()}}
\item \href{#method-VariableList-fromJSON}{\code{VariableList$fromJSON()}}
\item \href{#method-VariableList-toJSONString}{\code{VariableList$toJSONString()}}
\item \href{#method-VariableList-fromJSONString}{\code{VariableList$fromJSONString()}}
\item \href{#method-VariableList-validateJSON}{\code{VariableList$validateJSON()}}
\item \href{#method-VariableList-toString}{\code{VariableList$toString()}}
\item \href{#method-VariableList-isValid}{\code{VariableList$isValid()}}
\item \href{#method-VariableList-getInvalidFields}{\code{VariableList$getInvalidFields()}}
\item \href{#method-VariableList-print}{\code{VariableList$print()}}
\item \href{#method-VariableList-clone}{\code{VariableList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableList-new"></a>}}
\if{latex}{\out{\hypertarget{method-VariableList-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new VariableList class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableList$new(metadata = NULL, pagination = NULL, data = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{metadata}

\item{\code{pagination}}{pagination}

\item{\code{data}}{data}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-VariableList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableList-toList"></a>}}
\if{latex}{\out{\hypertarget{method-VariableList-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableList$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
VariableList as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of VariableList (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableList-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-VariableList-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert VariableList to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableList$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-VariableList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of VariableList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableList$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of VariableList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableList-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-VariableList-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableList$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
VariableList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableList-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-VariableList-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of VariableList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableList$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of VariableList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableList-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-VariableList-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to VariableList and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableList$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableList-toString"></a>}}
\if{latex}{\out{\hypertarget{method-VariableList-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableList$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of VariableList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableList-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-VariableList-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableList$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableList-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-VariableList-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableList$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableList-print"></a>}}
\if{latex}{\out{\hypertarget{method-VariableList-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableList$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VariableList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
