% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_list.R
\docType{class}
\name{VisitList}
\alias{VisitList}
\title{VisitList}
\format{
An \code{R6Class} generator object
}
\description{
VisitList Class

VisitList Class
}
\details{
Create a new VisitList
}
\examples{

## ------------------------------------------------
## Method `VisitList$toList`
## ------------------------------------------------

# convert array of VisitList (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{\link{Metadata} [optional]}

\item{\code{pagination}}{\link{Pagination} [optional]}

\item{\code{data}}{list(\link{Visit}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VisitList-new}{\code{VisitList$new()}}
\item \href{#method-VisitList-toJSON}{\code{VisitList$toJSON()}}
\item \href{#method-VisitList-toList}{\code{VisitList$toList()}}
\item \href{#method-VisitList-toSimpleType}{\code{VisitList$toSimpleType()}}
\item \href{#method-VisitList-fromJSON}{\code{VisitList$fromJSON()}}
\item \href{#method-VisitList-toJSONString}{\code{VisitList$toJSONString()}}
\item \href{#method-VisitList-fromJSONString}{\code{VisitList$fromJSONString()}}
\item \href{#method-VisitList-validateJSON}{\code{VisitList$validateJSON()}}
\item \href{#method-VisitList-toString}{\code{VisitList$toString()}}
\item \href{#method-VisitList-isValid}{\code{VisitList$isValid()}}
\item \href{#method-VisitList-getInvalidFields}{\code{VisitList$getInvalidFields()}}
\item \href{#method-VisitList-print}{\code{VisitList$print()}}
\item \href{#method-VisitList-clone}{\code{VisitList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisitList-new"></a>}}
\if{latex}{\out{\hypertarget{method-VisitList-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new VisitList class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisitList$new(metadata = NULL, pagination = NULL, data = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{metadata}

\item{\code{pagination}}{pagination}

\item{\code{data}}{data}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisitList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-VisitList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisitList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisitList-toList"></a>}}
\if{latex}{\out{\hypertarget{method-VisitList-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisitList$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
VisitList as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of VisitList (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisitList-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-VisitList-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert VisitList to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisitList$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisitList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-VisitList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of VisitList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisitList$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of VisitList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisitList-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-VisitList-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisitList$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
VisitList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisitList-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-VisitList-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of VisitList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisitList$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of VisitList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisitList-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-VisitList-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to VisitList and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisitList$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisitList-toString"></a>}}
\if{latex}{\out{\hypertarget{method-VisitList-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisitList$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of VisitList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisitList-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-VisitList-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisitList$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisitList-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-VisitList-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisitList$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisitList-print"></a>}}
\if{latex}{\out{\hypertarget{method-VisitList-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisitList$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisitList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VisitList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisitList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
